+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
+    xmlns:tools="http://schemas.android.com/tools">
 
     <application
         android:allowBackup="true"
@@ -11,26 +11,29 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Juda"
-        tools:targetApi="31" >
+        tools:targetApi="31">
         <activity
-            android:name=".FindMentorPostInfo.FindMentorPostInfo"
-            android:parentActivityName=".FindMentorList.FindMentorList"
+            android:name=".NewPost.FindMenteePost.NewFindMenteePost"
             android:exported="false" />
         <activity
-            android:name=".FindMentorList.FindMentorList"
-            android:parentActivityName=".MainActivity"
-            android:exported="false" />
+            android:name=".PostInfo.FindMentorPostInfo.FindMentorPostInfo"
+            android:exported="false"
+            android:parentActivityName=".PostList.FindMentorList.FindMentorList" />
         <activity
-            android:name=".FindMenteePostInfo.FindMenteePostInfo"
-            android:parentActivityName=".FindMenteeList.FindMenteeList"
-            android:exported="false" />
+            android:name=".PostList.FindMentorList.FindMentorList"
+            android:exported="false"
+            android:parentActivityName=".MainActivity" />
         <activity
-            android:name=".FindMenteeList.FindMenteeList"
-            android:parentActivityName=".MainActivity"
-            android:exported="false" />
+            android:name=".PostInfo.FindMenteePostInfo.FindMenteePostInfo"
+            android:exported="false"
+            android:parentActivityName=".PostList.FindMenteeList.FindMenteeList" />
+        <activity
+            android:name=".PostList.FindMenteeList.FindMenteeList"
+            android:exported="false"
+            android:parentActivityName=".MainActivity" />
         <activity
             android:name=".MainActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
