+++ app/src/main/java/com/example/juda/MainActivity.java
@@ -1,19 +1,15 @@
 package com.example.juda;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 
-import com.example.juda.FindMenteeList.FindMenteeList;
-import com.example.juda.FindMentorList.FindMentorList;
+import com.example.juda.PostList.FindMenteeList.FindMenteeList;
+import com.example.juda.PostList.FindMentorList.FindMentorList;
 
 public class MainActivity extends AppCompatActivity {
