+++ app/src/main/java/com/example/juda/NewPost/FindMenteePost/NewFindMenteePost.java
@@ -1,0 +1,16 @@
+package com.example.juda.NewPost.FindMenteePost;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.juda.R;
+
+public class NewFindMenteePost extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_new_find_mentee_post);
+    }
+}
