+++ app/src/main/java/com/example/juda/PostList/FindMenteeList/FindMenteeList.java
@@ -1,21 +1,21 @@
-package com.example.juda.FindMentorList;
+package com.example.juda.PostList.FindMenteeList;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import android.content.Intent;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.ListView;
 
-import com.example.juda.FindMenteePostInfo.FindMenteePostInfo;
-import com.example.juda.FindMentorPostInfo.FindMentorPostInfo;
+import com.example.juda.PostInfo.FindMenteePostInfo.FindMenteePostInfo;
 import com.example.juda.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -30,20 +30,23 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class FindMentorList extends AppCompatActivity {
+public class FindMenteeList extends AppCompatActivity {
+
 
     private Toolbar toolbar;
     final private FirebaseFirestore db = FirebaseFirestore.getInstance();
     private Button filter_BTN;
     private FloatingActionButton newPost_FAB;
-    private ListView findMentor_LV;
-    private FindMentorListAdapter mAdapter = null;
+    private ListView findMentee_LV;
+    private FindMenteeListAdapter mAdapter = null;
+
     private SimpleDateFormat format;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_find_mentor_list);
+        setContentView(R.layout.activity_find_mentee_list);
 
         Init();
     }
@@ -52,13 +55,13 @@ public class FindMentorList extends AppCompatActivity {
      * get ID and Connect
      */
     private void Init() {
-        toolbar = findViewById(R.id.basic_tool_bar_FindMentorList);
+        toolbar = findViewById(R.id.basic_tool_bar_FindMenteeList);
         toolbar.setTitle("");
         setSupportActionBar(toolbar);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        filter_BTN = findViewById(R.id.filter_BTN_FindMentorList);
-        newPost_FAB = findViewById(R.id.newPost_FAB_FindMentorList);
-        findMentor_LV = findViewById(R.id.findMentor_LV_FindMentorList);
+        filter_BTN = findViewById(R.id.filter_BTN_FindMenteeList);
+        newPost_FAB = findViewById(R.id.newPost_FAB_FindMenteeList);
+        findMentee_LV = findViewById(R.id.findMentee_LV_FindMenteeList);
 
         format = new SimpleDateFormat("yyyy. MM. dd");
 
@@ -68,23 +71,23 @@ public class FindMentorList extends AppCompatActivity {
 
     /**
      * This method gets data from firebase
-     * Get all data under 'MentorPost' and save all at menteePostHashMap
+     * Get all data under 'MenteePost' and save all at menteePostHashMap
      */
     private void getMenteePostList() {
-        db.collection("postRequest")
+        db.collection("postProvider")
                 .get()
                 .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                     @Override
                     public void onComplete(@NonNull Task<QuerySnapshot> task) {
                         if (task.isSuccessful()) {
-                            List<FindMentorPostData> dbData = new ArrayList<>();
+                            List<FindMenteePostData> dbData = new ArrayList<>();
                             for (QueryDocumentSnapshot document : task.getResult()) {
                                 Log.d("PIGMONGKEY", document.getId() + " => " + document.get("date"));
 
                                 Timestamp temp_timestamp = (Timestamp) document.get("date");
                                 Date temp_time = new Date(temp_timestamp.getSeconds()*1000);
 
-                                dbData.add(new FindMentorPostData(
+                                dbData.add(new FindMenteePostData(
                                         (String) document.getId(),
                                         (String) document.get("author"),
                                         (String) document.get("title"),
@@ -100,15 +103,15 @@ public class FindMentorList extends AppCompatActivity {
                 });
 
     }
-
+  
     /**
      * This method set onclick method to each item
      */
     private void setOnClick() {
-        findMentor_LV.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+        findMentee_LV.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(FindMentorList.this, FindMentorPostInfo.class);
+                Intent intent = new Intent(FindMenteeList.this, FindMenteePostInfo.class);
                 intent.putExtra("ID", mAdapter.getItem(position).getID());
                 startActivity(intent);
             }
@@ -116,12 +119,12 @@ public class FindMentorList extends AppCompatActivity {
     }
 
     /**
-     * Set adapter to findMentor_LV
-     * @param dbData - FindMentorPostData List
+     * Set adapter to findMentee_LV
+     * @param dbData - FindMenteePostData List
      */
-    private void setMenteePostListAdapter(List<FindMentorPostData> dbData) {
-        mAdapter = new FindMentorListAdapter(getApplicationContext(), dbData);
-        findMentor_LV.setAdapter(mAdapter);
+    private void setMenteePostListAdapter(List<FindMenteePostData> dbData) {
+        mAdapter = new FindMenteeListAdapter(getApplicationContext(), dbData);
+        findMentee_LV.setAdapter(mAdapter);
     }
 
     /**
