+++ app/src/main/java/com/example/juda/PostList/FindMenteeList/FindMenteeListAdapter.java
@@ -1,4 +1,8 @@
-package com.example.juda.FindMentorList;
+/*
+This adapter class is for findMentee_LV in FindMenteeList.java
+ */
+
+package com.example.juda.PostList.FindMenteeList;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,16 +11,15 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
 
-import com.example.juda.FindMenteeList.FindMenteePostData;
 import com.example.juda.R;
 
 import java.util.List;
 
-public class FindMentorListAdapter extends BaseAdapter {
+public class FindMenteeListAdapter extends BaseAdapter {
     private Context mContext = null;
-    List<FindMentorPostData> data;
+    List<FindMenteePostData> data;
 
-    public FindMentorListAdapter(Context mContext, List<FindMentorPostData> data) {
+    public FindMenteeListAdapter(Context mContext, List<FindMenteePostData> data) {
         this.mContext = mContext;
         this.data = data;
     }
@@ -27,7 +30,7 @@ public class FindMentorListAdapter extends BaseAdapter {
     }
 
     @Override
-    public FindMentorPostData getItem(int position) {
+    public FindMenteePostData getItem(int position) {
         return data.get(position);
     }
