+++ app/src/main/java/com/example/juda/PostList/FindMentorList/FindMentorList.java
@@ -1,21 +1,20 @@
-package com.example.juda.FindMenteeList;
+package com.example.juda.PostList.FindMentorList;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-
-import android.content.Intent;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.ListView;
 
-import com.example.juda.FindMenteePostInfo.FindMenteePostInfo;
+import com.example.juda.PostInfo.FindMentorPostInfo.FindMentorPostInfo;
 import com.example.juda.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -30,23 +29,20 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class FindMenteeList extends AppCompatActivity {
-
+public class FindMentorList extends AppCompatActivity {
 
     private Toolbar toolbar;
     final private FirebaseFirestore db = FirebaseFirestore.getInstance();
     private Button filter_BTN;
     private FloatingActionButton newPost_FAB;
-    private ListView findMentee_LV;
-    private FindMenteeListAdapter mAdapter = null;
-
+    private ListView findMentor_LV;
+    private FindMentorListAdapter mAdapter = null;
     private SimpleDateFormat format;
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_find_mentee_list);
+        setContentView(R.layout.activity_find_mentor_list);
 
         Init();
     }
@@ -55,13 +51,13 @@ public class FindMenteeList extends AppCompatActivity {
      * get ID and Connect
      */
     private void Init() {
-        toolbar = findViewById(R.id.basic_tool_bar_FindMenteeList);
+        toolbar = findViewById(R.id.basic_tool_bar_FindMentorList);
         toolbar.setTitle("");
         setSupportActionBar(toolbar);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        filter_BTN = findViewById(R.id.filter_BTN_FindMenteeList);
-        newPost_FAB = findViewById(R.id.newPost_FAB_FindMenteeList);
-        findMentee_LV = findViewById(R.id.findMentee_LV_FindMenteeList);
+        filter_BTN = findViewById(R.id.filter_BTN_FindMentorList);
+        newPost_FAB = findViewById(R.id.newPost_FAB_FindMentorList);
+        findMentor_LV = findViewById(R.id.findMentor_LV_FindMentorList);
 
         format = new SimpleDateFormat("yyyy. MM. dd");
 
@@ -71,23 +67,23 @@ public class FindMenteeList extends AppCompatActivity {
 
     /**
      * This method gets data from firebase
-     * Get all data under 'MenteePost' and save all at menteePostHashMap
+     * Get all data under 'MentorPost' and save all at menteePostHashMap
      */
     private void getMenteePostList() {
-        db.collection("postProvider")
+        db.collection("postRequest")
                 .get()
                 .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                     @Override
                     public void onComplete(@NonNull Task<QuerySnapshot> task) {
                         if (task.isSuccessful()) {
-                            List<FindMenteePostData> dbData = new ArrayList<>();
+                            List<FindMentorPostData> dbData = new ArrayList<>();
                             for (QueryDocumentSnapshot document : task.getResult()) {
                                 Log.d("PIGMONGKEY", document.getId() + " => " + document.get("date"));
 
                                 Timestamp temp_timestamp = (Timestamp) document.get("date");
                                 Date temp_time = new Date(temp_timestamp.getSeconds()*1000);
 
-                                dbData.add(new FindMenteePostData(
+                                dbData.add(new FindMentorPostData(
                                         (String) document.getId(),
                                         (String) document.get("author"),
                                         (String) document.get("title"),
@@ -103,15 +99,15 @@ public class FindMenteeList extends AppCompatActivity {
                 });
 
     }
-  
+
     /**
      * This method set onclick method to each item
      */
     private void setOnClick() {
-        findMentee_LV.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+        findMentor_LV.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(FindMenteeList.this, FindMenteePostInfo.class);
+                Intent intent = new Intent(FindMentorList.this, FindMentorPostInfo.class);
                 intent.putExtra("ID", mAdapter.getItem(position).getID());
                 startActivity(intent);
             }
@@ -119,12 +115,12 @@ public class FindMenteeList extends AppCompatActivity {
     }
 
     /**
-     * Set adapter to findMentee_LV
-     * @param dbData - FindMenteePostData List
+     * Set adapter to findMentor_LV
+     * @param dbData - FindMentorPostData List
      */
-    private void setMenteePostListAdapter(List<FindMenteePostData> dbData) {
-        mAdapter = new FindMenteeListAdapter(getApplicationContext(), dbData);
-        findMentee_LV.setAdapter(mAdapter);
+    private void setMenteePostListAdapter(List<FindMentorPostData> dbData) {
+        mAdapter = new FindMentorListAdapter(getApplicationContext(), dbData);
+        findMentor_LV.setAdapter(mAdapter);
     }
 
     /**
