+++ app/src/main/res/layout/activity_find_mentee_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FindMenteeList.FindMenteeList">
+    tools:context=".PostList.FindMenteeList.FindMenteeList">
 
     <include
         android:id="@+id/basic_tool_bar_FindMenteeList"
